docs/encoders.md: recommend a pull-up resistor on CS pin for AMS encoders
In my experience when attempting the initial configuration and
calibration (i.e. after `erase_configuration()`) of an AS5048A
encoder, the ODrive invariably ends up being stuck in an
`ENCODER_ERROR_ABS_SPI_COM_FAIL` state.

After `erase_configuration()` most of the GPIO pins are low, either
because they are initialized as `GPIO_MODE_DIGITAL`, which is low by
default, or as `GPIO_MODE_ANALOG_IN`, which tends to float quite close
to 0V.  It takes some time to configure one or two GPIOs as
`abs_spi_cs_gpio_pin`, save the configuration, and restart the ODrive,
especially when one copy-pastes or, worse, types the configuration
commands into `odrivetool`.  All this while those GPIO pins remain
low, and if an encoder is connected to one of them, then it will
interpret that as active Chip Select.  During this time, however, the
ODrive does send some commands to the DRV8031 chips through the shared
CLK/MOSI/MISO lines.  The encoder sees these commands, and, because of
the active CS, it attempts to interpret them, gets confused, and sets
the error flag in each subsequent packet, and continues to do so until
the error is cleared (what, alas, the ODrive never does).

This issue can be entirely avoided by an external pull-up resistor on
the CS pin, because it ensures that the CS only ever goes low when the
ODrive does want to talk to that specific encoder.

Add a note to `docs/encoders.md` recommending an external pull-up on
the CS pin of AMS encoders, in the hope that it will save other users
some trouble.

CUI AMT23xx encoders don't need a pull-up on their CS, because their
protocol doesn't even have an error flag in the first place.  I
haven't checked whether other SPI encoders would benefit from a
pull-up on CS, but I guess it can't hurt any of them.